+++ previous-versions/commerce/RELEASES.md
@@ -341,3 +341,13 @@ Based on:
 - [go v0.25.0] previous-versions/commerce
 ### Releases
 - [Go v0.25.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/commerce/v0.25.0 - previous-versions/commerce
+
+## 2023-11-06 01:29:15
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.26.0] previous-versions/commerce
+### Releases
+- [Go v0.26.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/commerce/v0.26.0 - previous-versions/commerce